updated unit test for moving average

+++ stock-price-data-model/src/main/java/space/gavinklfong/demo/kafka/model/CountAndSum.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @NoArgsConstructor(force = true)
 @AllArgsConstructor
 @Value
 public class CountAndSum {
     Long count;
-    Double sum;
+    BigDecimal sum;
 }
